+++ src/main/java/com/telepathicgrunt/blame/main/MissingNBTBlame.java
@@ -44,8 +44,8 @@ public class MissingNBTBlame {
 				"\n\n Failed to load structure nbt file from:  " + miniRL +
 				"\n The path represented by this is:  " + fullPath +
 				(parentID != null ? "\n The calling Template Pool is: " + parentID : "") +
-				"\n Most common cause is that the structure nbt file is not actually at that location." +
-				"\n Please let the mod author know about this so they can move their structure nbt file to the correct place." +
+				"\n Most common cause is that there is a typo in this path to the nbt file." +
+				"\n Please let the mod author know about this so they can double check to make sure the path is correct." +
 				"\n A common mistake is putting the structure nbt file in the asset folder when it needs to go in data/structures folder.\n";
 		Blame.LOGGER.log(Level.ERROR, errorReport);